Re #896 correct binning for single pixels

+++ horace_core/sqw/axes_blocks/@AxesBlockBase/private/bin_pixels_.m
@@ -185,7 +185,7 @@ if ndims == 0
     end
 else
     for i=1:ndata
-        out{i} = out{i}+accumarray(pix_indx,bin_values{i}(ok),size(npix));
+        out{i} = out{i}+accumarray(pix_indx,double(bin_values{i}(ok)),size(npix));
     end
 end
 s = out{1};
